requirements.txt
aiohttp==3.8.6
wom-py==0.8.1
requests==2.31.0
